Added spinbox for setting the number of games to sim
Added spinbox to play controls for setting the number of games to sim when clicking the double skip button.
Added LCD type box to the consolations row, displaying the total number of games simulated.
added forth data display form column and set the forms to be equa spaced along the bottom.